backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.14.0
# psycopg2-binary - PostgreSQL uchun (ixtiyoriy, faqat production uchun)
# psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv==1.0.1
email-validator==2.2.0
slowapi==0.1.9
httpx==0.27.0
# twilio==9.0.0  # Twilio uchun (ixtiyoriy)

